Implemented GET and POST methods in Plants and GET method in PlantByID
Plants class - The GET method retrieves all plants from the database and returns them as a list of dictionaries. The POST method allows for the creation of new plants in the database using data received in the request.

PlantByID class - The GET method retrieves a specific plant by its ID from the database.